moving postgres rest api into the app folder
