+++ app/pg_rest_api/Procfile
@@ -1,0 +1,2 @@
+pip uninstall syft --y
+web: gunicorn app:app --log-file -
