--- grid/rest_api/__init__.py
@@ -1,1 +1,0 @@
-from flask import Flask
