+++ test/flask_base_test.py
@@ -1,6 +1,6 @@
 import unittest
-from grid.app import create_app, db
-from grid.app.models import Worker, WorkerObject
+from app.pg_rest_api import create_app, db
+from app.pg_rest_api.models import Worker, WorkerObject
 
 
 class FlaskBaseTestCase(unittest.TestCase):
